refactor(sdk): replace socket.io with native ws transport
